refactor: decorate bulk action
