+++ packages/app-page-builder/src/admin/components/BulkActions/ActionPublish.tsx
@@ -1,28 +1,19 @@
 import React, { useMemo } from "react";
 import { ReactComponent as PublishIcon } from "@material-design-icons/svg/outlined/publish.svg";
-import { useFolders, useRecords } from "@webiny/app-aco";
+import { useRecords } from "@webiny/app-aco";
 import { observer } from "mobx-react-lite";
 import { PageListConfig } from "~/admin/config/pages";
 import { useAdminPageBuilder } from "~/admin/hooks/useAdminPageBuilder";
-import { usePagesPermissions } from "~/hooks/permissions";
 import { getPagesLabel } from "~/admin/components/BulkActions/BulkActions";
 
 export const ActionPublish = observer(() => {
-    const { canPublish } = usePagesPermissions();
     const { publishPage, client } = useAdminPageBuilder();
     const { getRecord } = useRecords();
-    const { folderLevelPermissions: flp } = useFolders();
 
     const { useWorker, useButtons, useDialog } = PageListConfig.Browser.BulkAction;
     const { IconButton } = useButtons();
     const worker = useWorker();
     const { showConfirmationDialog, showResultsDialog } = useDialog();
-
-    const canPublishAll = useMemo(() => {
-        return worker.items.every(item => {
-            return canPublish() && flp.canManageContent(item.location?.folderId);
-        });
-    }, [worker.items]);
 
     const pagesLabel = useMemo(() => {
         return getPagesLabel(worker.items.length);
@@ -74,11 +65,6 @@ export const ActionPublish = observer(() => {
                 });
             }
         });
-
-    if (!canPublishAll) {
-        console.log("You don't have permissions to publish pages.");
-        return null;
-    }
 
     return (
         <IconButton
