+++ packages/app-page-builder/src/admin/views/Pages/PagesModule.tsx
@@ -7,7 +7,7 @@ import {
     ActionMove,
     ActionPublish,
     ActionUnpublish
-} from "~/admin/components/BulkActions";
+} from "./BulkActions";
 import { DeleteFolder, EditFolder, SetFolderPermissions } from "@webiny/app-aco";
 import {
     CellActions,
